requirements.txt
streamlit==0.66.0
pandas==1.1.4
matplotlib==3.3.3
plotly.express==4.10.0
